requirements.txt
gradio==4.28.3
transformers==4.41.2
torch
googletrans==4.0.0-rc1
PyPDF2==3.0.1
python-docx==1.1.2